Refactoring FileUtils
